move SBOM conversion code to own package
