+++ pkg/collector/corechecks/sbom/processor.go
@@ -24,6 +24,7 @@ import (
 	pkgconfigsetup "github.com/DataDog/datadog-agent/pkg/config/setup"
 
 	"github.com/DataDog/datadog-agent/pkg/sbom"
+	"github.com/DataDog/datadog-agent/pkg/sbom/bomconvert"
 	"github.com/DataDog/datadog-agent/pkg/sbom/collectors/host"
 	"github.com/DataDog/datadog-agent/pkg/sbom/collectors/procfs"
 	sbomscanner "github.com/DataDog/datadog-agent/pkg/sbom/scanner"
@@ -238,7 +239,7 @@ func (p *processor) processHostScanResult(result sbom.ScanResult) {
 		Id:                 p.hostname,
 		InUse:              true,
 		GeneratedAt:        timestamppb.New(result.CreatedAt),
-		GenerationDuration: convertDuration(result.Duration),
+		GenerationDuration: bomconvert.ConvertDuration(result.Duration),
 	}
 
 	if result.Error != nil {
@@ -262,7 +263,7 @@ func (p *processor) processHostScanResult(result sbom.ScanResult) {
 				sbom.Status = model.SBOMStatus_FAILED
 			} else {
 				sbom.Sbom = &model.SBOMEntity_Cyclonedx{
-					Cyclonedx: convertBOM(report),
+					Cyclonedx: bomconvert.ConvertBOM(report),
 				}
 			}
 
@@ -306,7 +307,7 @@ func (p *processor) processProcfsScanResult(result sbom.ScanResult) {
 		Type:               model.SBOMSourceType_CONTAINER_FILE_SYSTEM,
 		InUse:              true,
 		GeneratedAt:        timestamppb.New(result.CreatedAt),
-		GenerationDuration: convertDuration(result.Duration),
+		GenerationDuration: bomconvert.ConvertDuration(result.Duration),
 	}
 
 	if result.Error != nil {
@@ -333,7 +334,7 @@ func (p *processor) processProcfsScanResult(result sbom.ScanResult) {
 				sbom.Status = model.SBOMStatus_FAILED
 			} else {
 				sbom.Sbom = &model.SBOMEntity_Cyclonedx{
-					Cyclonedx: convertBOM(report),
+					Cyclonedx: bomconvert.ConvertBOM(report),
 				}
 			}
 		}
@@ -453,9 +454,9 @@ func (p *processor) processImageSBOM(img *workloadmeta.ContainerImageMetadata) {
 		default:
 			sbom.Status = model.SBOMStatus_SUCCESS
 			sbom.GeneratedAt = timestamppb.New(img.SBOM.GenerationTime)
-			sbom.GenerationDuration = convertDuration(img.SBOM.GenerationDuration)
+			sbom.GenerationDuration = bomconvert.ConvertDuration(img.SBOM.GenerationDuration)
 			sbom.Sbom = &model.SBOMEntity_Cyclonedx{
-				Cyclonedx: convertBOM(img.SBOM.CycloneDXBOM),
+				Cyclonedx: bomconvert.ConvertBOM(img.SBOM.CycloneDXBOM),
 			}
 		}
 		p.queue <- sbom
